Integrated uploadScreen into HomeView Reworked uploadView
