+++ pb-app/app/src/main/java/com/github/eeriefoods/pizzabrei/MainActivity.kt
@@ -5,6 +5,7 @@ package com.github.eeriefoods.pizzabrei
 import android.Manifest.permission.WRITE_EXTERNAL_STORAGE
 import android.content.pm.PackageManager
 import android.os.Bundle
+import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -62,6 +63,7 @@ class MainActivity : ComponentActivity() {
             )
         )
         super.onCreate(savedInstanceState)
+        window.setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE);
         setContent {
             PizzaBreiTheme {
                 Surface(
