+++ pb-app/app/src/main/java/com/github/eeriefoods/pizzabrei/presentation/ui/navigation/NavGraph.kt
@@ -12,7 +12,7 @@ import com.github.eeriefoods.pizzabrei.presentation.ui.views.home.HomeViewModel
 import com.github.eeriefoods.pizzabrei.presentation.ui.views.home.HomeScreen
 import com.github.eeriefoods.pizzabrei.presentation.ui.views.search.SearchScreen
 import com.github.eeriefoods.pizzabrei.presentation.ui.views.settings.SettingsScreen
-import com.github.eeriefoods.pizzabrei.presentation.ui.views.upload.uploadScreen
+import com.github.eeriefoods.pizzabrei.presentation.ui.views.upload.uploadView
 
 
 @ExperimentalFoundationApi
@@ -38,7 +38,7 @@ fun NavGraph(
             SearchScreen(navController)
         }
         composable(route = Screens.Upload.route){
-            uploadScreen(navController, activity)
+            uploadView(navController, activity)
         }
     }
 }
