+++ pb-app/app/src/main/java/com/github/eeriefoods/pizzabrei/presentation/ui/views/upload/UploadViewModel.kt
@@ -2,8 +2,6 @@ package com.github.eeriefoods.pizzabrei.presentation.ui.views.upload
 
 
 
-class UploadScreenModel {
-
-
+class UploadViewModel {
 
 }
